Separation of model and networks file for scalabillity
